requirements.txt
autogen-agentchat
autogen-ext[openai]
autogen-core
playwright
aiohttp
aiofiles
openai
python-dotenv
Scrapy
knockpy
sqlmap
wapiti3
arjun
requests
#knockpy might give an error, but it is still installed
